Initial file push.
As you may notice, I have left framework detection and previously used framework functions in on the server side, this is because the majority of my releases will be framework driven so I thought I may as well leave this in as someone else may also be using a framework. However, if this isn't the case, all you have to do is simply delete the framework file on both the client and server side and remove any functions that use framework. If you're confident enough to be making your own resources, this shouldn't be a problem at all but if you're struggling to do this you can contact me in my community Discord found on my GitHub profile.